# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "JobTrack API",
    "version": "1.0.0",
    "description": "JobTrack backend API documentation"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/api/applications": {
      "get": {
        "summary": "List applications",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "stage",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/StageType"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Applications list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "applications"
                  ],
                  "properties": {
                    "applications": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Application"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create application",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created application",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Application"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/applications/{id}": {
      "get": {
        "summary": "Get application detail",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ApplicationId"
          }
        ],
        "responses": {
          "200": {
            "description": "Application detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "summary": "Update application",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ApplicationId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApplicationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated application",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Application"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "Delete application",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ApplicationId"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/applications/{id}/events": {
      "get": {
        "summary": "List events for application",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ApplicationId"
          }
        ],
        "responses": {
          "200": {
            "description": "Events list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "post": {
        "summary": "Create event for application",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ApplicationId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created event",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/applications/{id}/documents": {
      "post": {
        "summary": "Upload PDF document",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ApplicationId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Uploaded document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "400": {
            "description": "Invalid file"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/documents/{id}": {
      "delete": {
        "summary": "Delete document",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DocumentId"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jobtrack_auth"
      }
    },
    "parameters": {
      "ApplicationId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "DocumentId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "schemas": {
      "CareerType": {
        "type": "string",
        "enum": [
          "new",
          "experienced",
          "any"
        ]
      },
      "StageType": {
        "type": "string",
        "enum": [
          "interest",
          "applied",
          "document_pass",
          "assignment",
          "interview",
          "final_pass",
          "rejected"
        ]
      },
      "SourceType": {
        "type": "string",
        "enum": [
          "saramin",
          "jobkorea",
          "company",
          "linkedin",
          "etc"
        ]
      },
      "EventType": {
        "type": "string",
        "enum": [
          "deadline",
          "result",
          "interview"
        ]
      },
...
      "Application": {
        "type": "object",
        "required": [
          "id",
          "user_id",
          "company_name",
          "position",
          "career_type",
          "current_stage"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "company_name": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "career_type": {
            "$ref": "#/components/schemas/CareerType"
          },
          "job_url": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SourceType"
              }
            ],
            "nullable": true
          },
          "merit_tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "current_stage": {
            "$ref": "#/components/schemas/StageType"
          },
          "company_memo": {
            "type": "string",
            "nullable": true
          },
          "cover_letter": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "CreateApplicationRequest": {
        "type": "object",
        "required": [
          "company_name",
          "position",
          "career_type"
        ],
        "properties": {
          "company_name": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "career_type": {
            "$ref": "#/components/schemas/CareerType"
          },
          "job_url": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SourceType"
              }
            ],
            "nullable": true
          },
          "merit_tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "current_stage": {
            "$ref": "#/components/schemas/StageType"
          },
          "company_memo": {
            "type": "string",
            "nullable": true
          },
          "cover_letter": {
            "type": "string",
            "nullable": true
          },
          "deadline": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UpdateApplicationRequest": {
        "type": "object",
        "properties": {
          "company_name": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "career_type": {
            "$ref": "#/components/schemas/CareerType"
          },
          "job_url": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SourceType"
              }
            ],
            "nullable": true
          },
          "merit_tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "current_stage": {
            "$ref": "#/components/schemas/StageType"
          },
          "company_memo": {
            "type": "string",
            "nullable": true
          },
          "cover_letter": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Event": {
        "type": "object",
        "required": [
          "id",
          "application_id",
          "event_type",
          "scheduled_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "application_id": {
            "type": "string",
            "format": "uuid"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventType"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "interview_round": {
            "type": "integer",
            "nullable": true
          },
          "notified_d3": {
            "type": "boolean",
            "nullable": true
          },
          "notified_d1": {
            "type": "boolean",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "CreateEventRequest": {
        "type": "object",
        "required": [
          "event_type",
          "scheduled_at"
        ],
        "properties": {
          "event_type": {
            "$ref": "#/components/schemas/EventType"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "interview_round": {
            "type": "integer",
            "nullable": true
          }
        }
      },
...
      "Document": {
        "type": "object",
        "required": [
          "id",
          "application_id",
          "file_name",
          "file_size",
          "file_url"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "application_id": {
            "type": "string",
            "format": "uuid"
          },
          "file_name": {
            "type": "string"
          },
          "file_size": {
            "type": "integer"
          },
          "file_url": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "ApplicationDetailResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Application"
          },
          {
            "type": "object",
            "required": [
              "events",
              "documents"
            ],
            "properties": {
              "events": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Event"
                }
              },
              "documents": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          }
        ]
      }
    }
  }
}